third_party/WebKit/LayoutTests/http/tests/serviceworker/postmessage.html

<!DOCTYPE html>
<title>Service Worker: postMessage</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="resources/test-helpers.js"></script>
<script>
var test = async_test('postMessage to a ServiceWorker (and back via MessagePort)');
test.step(function() {
    var scope = 'resources/blank.html';
    navigator.serviceWorker.unregister(scope).then(
        doTest,
        unreached_rejection(test, 'Unregister should not fail')
    );

    function doTest() {
      navigator.serviceWorker.register('resources/postmessage-worker.js',
                                       {scope: scope}).then(
          test.step_func(function(worker) {
              var messageChannel = new MessageChannel();
              messageChannel.port1.onmessage = test.step_func(onMessage);

              worker.postMessage({port: messageChannel.port2}, [messageChannel.port2]);

              worker.postMessage({value: 1});
              worker.postMessage({value: 2});
              worker.postMessage({done: true});
          }),
          unreached_rejection(test, 'Registration should succeed, but failed'));
    };

    var result = [];
    var expected = [
        'Acking value: 1',
        'Acking value: 2',
    ];

    function onMessage(e) {
        var message = e.data;
        if (message === 'quit') {
            assert_array_equals(result, expected, 'Worker should post back expected values.');
            test.done();
        } else {
            result.push(message);
        }
    };
});
</script>
